error handling for save, delete and update (#6)
added error messages when an id is not positive, a message is empty or an id cannot be found. Added imports for mock classes to avoid errors

+++ src/main/java/com/example/budgettrackerv1/controller/IncomeController.java
@@ -1,11 +1,14 @@
 package com.example.budgettrackerv1.controller;
 
 import com.example.budgettrackerv1.Constants;
+import com.example.budgettrackerv1.MockData;
+import com.example.budgettrackerv1.model.Category;
 import com.example.budgettrackerv1.model.Income;
 import com.example.budgettrackerv1.service.IncomeService;
 import com.google.gson.Gson;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -92,6 +95,10 @@ public class IncomeController {
 
     @PostMapping("/saveIncome")
     public ResponseEntity<String> save(@RequestBody String jsonIncome) {
+        if(jsonIncome == null || jsonIncome.isEmpty()) {
+            String message = "Income cannot be empty";
+            return ResponseEntity.badRequest().body(message);
+        }
         Income income = this.gson.fromJson(jsonIncome, Income.class);
         System.out.println(income);
         this.INCOME_SERVICE.save(income);
@@ -102,8 +109,39 @@ public class IncomeController {
 
     @DeleteMapping("/deleteIncome/{id}")
     public ResponseEntity<String> delete(@PathVariable int id) {
-        this.INCOME_SERVICE.delete(id);
+        if (id <= 0) {
+            String message = "Income ID must be a positive integer";
+            return ResponseEntity.badRequest().body(message);
+        }
+        try {
+            this.INCOME_SERVICE.delete(id);
+        } catch (Exception e) {
+            String message = String.format("Income with id %d could not be found", id);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(message);
+        }
         String message = String.format("Income with id %d was deleted successfully", id);
+        return ResponseEntity.ok(message);
+    }
+
+    @PutMapping("/updateIncome/{id}")
+    public ResponseEntity<String> update(@RequestBody String jsonIncome, @PathVariable int id) {
+        if(jsonIncome == null || jsonIncome.isEmpty()) {
+            String message = "Income cannot be empty";
+            return ResponseEntity.badRequest().body(message);
+        } else if (id <= 0) {
+            String message = "Income ID must be a positive integer";
+            return ResponseEntity.badRequest().body(message);
+        }
+        Income income = this.gson.fromJson(jsonIncome, Income.class);
+        income.setId(id);
+        System.out.println(income);
+        try {
+            this.INCOME_SERVICE.update(income);
+        } catch (Exception e) {
+            String message = String.format("Income with id %d could not be found", id);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(message);
+        }
+        String message = String.format("Income with id %d was updated successfully", id);
         return ResponseEntity.ok(message);
     }
 }
